pyproject.toml
[project]
name = "mega-meem"
version = "0.1.0"
description = "Future Asif — a personal AI chat backend"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.77.0",
    "fastapi>=0.128.0",
    "google-genai>=1.62.0",
    "pydantic-settings>=2.12.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
